--- a/src/Frontend/Components/AppContainer/__tests__/AppContainer.test.tsx
+++ b/src/Frontend/Components/AppContainer/__tests__/AppContainer.test.tsx
@@ -10,6 +10,8 @@ import { IpcRenderer } from 'electron';
 
 let originalIpcRenderer: IpcRenderer;
 
+jest.mock('../../ResourceDetailsTabs/get-new-accordion-worker');
+
 describe('The AppContainer', () => {
   beforeAll(() => {
     originalIpcRenderer = global.window.ipcRenderer;