--- a/src/Frontend/Components/AuditView/__tests__/AuditView.test.tsx
+++ b/src/Frontend/Components/AuditView/__tests__/AuditView.test.tsx
@@ -7,6 +7,8 @@ import { AuditView } from '../AuditView';
 import React from 'react';
 import { renderComponentWithStore } from '../../../test-helpers/render-component-with-store';
 
+jest.mock('../../ResourceDetailsTabs/get-new-accordion-worker');
+
 describe('The AuditView', () => {
   test('renders AuditView', () => {
     renderComponentWithStore(<AuditView />);