--- a/src/Frontend/Components/ResourceDetailsTabs/__mocks__/get-new-accordion-worker.ts
+++ b/src/Frontend/Components/ResourceDetailsTabs/__mocks__/get-new-accordion-worker.ts
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: Facebook, Inc. and its affiliates
+// SPDX-FileCopyrightText: TNG Technology Consulting GmbH <https://www.tngtech.com>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+export function getNewAccordionWorker(): Worker {
+  throw new Error('JEST-MOCK-GET-NEW-ACCORDION-WORKER');
+}