--- a/src/Frontend/integration-tests/all-views-tests/__tests-ci__/open_file_in_external_link.test.tsx
+++ b/src/Frontend/integration-tests/all-views-tests/__tests-ci__/open_file_in_external_link.test.tsx
@@ -24,6 +24,8 @@ import { clickOnElementInResourceBrowser } from '../../../test-helpers/resource-
 
 let originalIpcRenderer: IpcRenderer;
 
+jest.mock('../../../Components/ResourceDetailsTabs/get-new-accordion-worker');
+
 jest.setTimeout(TEST_TIMEOUT);
 
 function mockElectronBackend(mockChannelReturn: ParsedFileContent): void {