--- a/src/Frontend/integration-tests/all-views-tests/__tests-ci__/other-popups.test.tsx
+++ b/src/Frontend/integration-tests/all-views-tests/__tests-ci__/other-popups.test.tsx
@@ -50,6 +50,8 @@ import { ATTRIBUTION_SOURCES } from '../../../../shared/shared-constants';
 
 let originalIpcRenderer: IpcRenderer;
 
+jest.mock('../../../Components/ResourceDetailsTabs/get-new-accordion-worker');
+
 jest.setTimeout(TEST_TIMEOUT);
 
 function mockElectronBackend(mockChannelReturn: ParsedFileContent): void {