--- a/src/Frontend/integration-tests/audit-view-tests/__tests-ci__/add-to-attribution.test.tsx
+++ b/src/Frontend/integration-tests/audit-view-tests/__tests-ci__/add-to-attribution.test.tsx
@@ -40,6 +40,8 @@ import { clickOnElementInResourceBrowser } from '../../../test-helpers/resource-
 
 let originalIpcRenderer: IpcRenderer;
 
+jest.mock('../../../Components/ResourceDetailsTabs/get-new-accordion-worker');
+
 jest.setTimeout(TEST_TIMEOUT);
 
 function mockElectronBackend(mockChannelReturn: ParsedFileContent): void {