### Summary of changes

Introduce web workers and compute accordion data in a worker (async, not in the main process). Since Webpack 5 web workers are supported by default, not requiring any external package or change in the Webpack configuration. Further changes:
- Web worker can run out of memory with big files, such error is caught and the computation is then performed as fallback on main,
- jest does not support web workers yet, therefore they need to be mocked in the test. The mock just throws an error in the worker and triggers the fallback computation on main.

### Context and reason for change

The app gets stuck for a while when clicking on a folder with a lot of signals inside. Web workers allow to calculate the accordion faster, meanwhile leaving the main process free to process user input.

### How can the changes be tested

Open keycloak input and navigate away from the root folder (somewhere where there is an attribution). Then click on the root folder. On the right and above the accordion the UI will change instantly, meanwhile the old accordion will be shown until the new one is ready.
The fallback path can be tested using the ORT input file.