@benedikt-richter 2022-01-25
Just did a quick performance test with the build version. Opening the root folder of a huge app:
Before: 800ms (total) = approx 1s
Now: 260ms (to display the root folders attributions) + 1100ms (load package panels) + approx 700ms (idle) = approx 2s